% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{marginal}
\alias{marginal}
\title{Predict the the marginal response surface wrt Z using the Gaussian process fit.}
\usage{
marginal(object, newX, newZ, causal = FALSE)
}
\arguments{
\item{object}{An "ace" object from the GPspline function}

\item{newX}{A matrix with new data. If not presented, using the training data.}

\item{newZ}{A vector, matrix or scalar with new treatment data. If it is a scalar, it predicts using the same value for all observations. If missing, it uses the training data.}

\item{causal}{A logical}

\item{marginal}{A logical}
}
\value{
Returns the MAP and the 95 percent credible interval of the fitted process as a list.
}
\description{
Predict the the marginal response surface wrt Z using the Gaussian process fit.
}
