% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GPspline.R
\name{plot.ace}
\alias{plot.ace}
\title{Plot marginal effect.}
\usage{
\method{plot}{ace}(object, Xcol, marginal = FALSE, plot3D = FALSE,
  show.observations = TRUE, Xlim = c(-1, 1), Xstep = 0.1, Zlim = c(-1,
  1), Zstep = 0.1, truefun)
}
\arguments{
\item{object}{An "GPspline" object from the GPspline function}

\item{Xcol}{Either numeric or character, the latter only works if the training was done with a data.frame (i.e. not GPspline.train). If missing plots the 2D graph of Y on Z.}

\item{marginal}{A logical statement, default: \code{FALSE}. If \code{TRUE}, plots the marginal surface of the response surface Y with respect to the basis expanded model Z using average X.}

\item{plot3D}{A logical statemnt whether a three-dimensional plot should be produced. This is only valid if an X-column is selected and the plotly package is installed.}

\item{show.observations}{A logical}

\item{Xlim}{A two-dimensional vector}

\item{Xstep}{A scalar}

\item{Zlim}{A two-dimensional vector}

\item{Zstep}{A scalar}

\item{truefun}{A function handle for simulations.}
}
\value{
Returns the plotly object
}
\description{
Plot marginal effect.
}
