% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GPspline.R
\name{plot.GPspline}
\alias{plot.GPspline}
\title{Plot marginal effect.}
\usage{
\method{plot}{GPspline}(object, Xcol, marginal = TRUE, Xlim = c(-1, 1),
  Xstep = 0.1, Zlim = c(-1, 1), Zstep = 0.1, truefun)
}
\arguments{
\item{object}{An "GPspline" object from the GPspline function}

\item{Xcol}{Either numeric or character, the latter only works if the training was done with a data.frame (i.e. not GPspline.train). If missing plots the 2D graph of Y on Z.}

\item{marginal}{A logical statement, default: \code{FALSE}. If \code{TRUE}, plots the marginal surface of the response surface Y with respect to the basis expanded model Z using average X.}

\item{Xlim}{A logical}

\item{Xstep}{A logical}

\item{Zlim}{A logical}

\item{Zstep}{A logical}

\item{truefun}{For simulation tests}
}
\value{
Returns the plotly object
}
\description{
Plot marginal effect.
}
