% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ace.R
\name{plot_ace}
\alias{plot_ace}
\title{Plot predictions of the Additive Causal Expansion model}
\usage{
plot_ace(object, Xcol, marginal = FALSE, plot3D = FALSE,
  show.observations = TRUE, Xlim = c(-1, 1), Xstep = 0.1, Zlim = c(-1,
  1), Zstep = 0.1, truefun, ...)
}
\arguments{
\item{object}{An \code{ace} object as generated by the ace.train function.}

\item{Xcol}{A numeric. If missing plots the 2D graph of Y on Z.}

\item{marginal}{A logical statement, default: \code{FALSE}. If \code{TRUE}, plots the marginal surface of the response surface Y with respect to the basis expanded model Z using average X.}

\item{plot3D}{A logical statement whether a three-dimensional plot should be produced. This is only valid if an X-column is selected and the plotly package is installed.}

\item{show.observations}{A logical scalar that determines}

\item{Xlim}{A two-dimensional vector with thelimits of the confounder dimension.}

\item{Xstep}{A scalar for the step-size of the confounder dimension.}

\item{Zlim}{A two-dimensional vector with thelimits of the treatment dimension.}

\item{Zstep}{A scalar for the step-size of the treatment dimension.}

\item{truefun}{A function handle for simulations.}

\item{...}{Arguments handed to plotting function (\code{plot}, \code{contour}, \code{plotly}).}
}
\value{
Returns the plot object
}
\description{
This function plots the response surface or the marginal effect fitted by the \code{ace} package around the training sample. For non-binary treatments, the method can display the interaction of the treatment with one confounder.
The package plots:
\itemize{
\item for binary treatments, a 2D plot with \code{X[, Xcol]} on the x-axis. This plots,
\itemize{
\item if \code{marginal = FALSE}: the treatment and control reponse surfaces.
\item if \code{marginal = TRUE}: the treatment effect.
}
\item For non-binary (continuous and -for now- categorial) treatments the method can display a 2D contour or a 3D surface plot with the interaction between the treatment and a specified confounder through \code{Xcol}.
If no confounder is selected, a simple 2D plot with the treatment on the x-axis is returned. Depending on \code{marginal} the levels correspond either to the response surface (if \code{FALSE}, default) or the marginal treatment effect/slope of response surface (if \code{TRUE}).
The different confounder plots can be accessed using:
\itemize{
\item if \code{plot3D = TRUE} and \code{Xcol} specified: A 3D surface using plotly (if \code{plotly} installed, otherwise plot3D defaults to FALSE). The treatment variable and the selected \code{X[, Xcol]} are shown on the x- and y-axis, the outcome surface on the z-axis.
\item if \code{plot3D = FALSE} and \code{Xcol} specified: the treatment effect for different treatment levels using a (2D) contour plot with the treatment variable and the selected \code{X[, Xcol]} on the x- and y-axis.
\item if no \code{Xcol} specified: A 2D plot with treatment levels on the x- and the outcome on the y-axis.
}
}
}
\details{
Note: The method is using a colorblind-robust palette as in http://jfly.iam.u-tokyo.ac.jp/color/#what.
}
