% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GPspline.R
\name{plot.GPspline}
\alias{plot.GPspline}
\title{Plot marginal effect.}
\usage{
\method{plot}{GPspline}(object, marginal = TRUE, plotly = TRUE,
  Xlim = c(-1, 1), Xstep = 0.1, Zlim = c(-1, 1), Zstep = 0.1)
}
\arguments{
\item{object}{An "GPspline" object from the GPspline function}

\item{marginal}{A matrix with new data. If not presented, using the training data.}

\item{plotly}{A vector, matrix or scalar with new treatment data. If it is a scalar, it predicts using the same value for all observations. If missing, it uses the training data.}

\item{Xlim}{A logical}

\item{Xstep}{A logical}

\item{Zlim}{A logical}

\item{Zstep}{A logical}
}
\value{
Returns the plotly object
}
\description{
Plot marginal effect.
}
