% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ace.R
\name{predict.ace}
\alias{predict.ace}
\title{Predict the the response surface and the marginal response wrt Z using the Gaussian process fit.}
\usage{
\method{predict}{ace}(object, newX, newZ, marginal = FALSE, causal = FALSE,
  normalize = TRUE)
}
\arguments{
\item{object}{An "ace" object from the ace}

\item{newX}{A matrix with new data. If not presented, using the training data.}

\item{newZ}{A vector, matrix or scalar with new treatment data. If it is a scalar, it predicts using the same value for all observations. If missing, it uses the training data.}

\item{marginal}{A logical}

\item{causal}{logical}
}
\value{
Returns the MAP and the 95 percent credible interval of the fitted process as a list.
}
\description{
Predict the the response surface and the marginal response wrt Z using the Gaussian process fit.
}
