% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_treatment.R
\name{robust_treatment}
\alias{robust_treatment}
\title{Calculate robust average treatment effect using posterior uncertainty}
\usage{
robust_treatment(object, newX, n.steps = 10)
}
\arguments{
\item{X}{The covariates used in the causal regression.}
}
\value{
A trained causal forest object.
}
\description{
Calculate robust average treatment effect using posterior uncertainty
}
\examples{
\dontrun{

}

}
